Fix GPIO pin random value
The value on the GPIO pin is used to enter bootloader mode.
If the value is low, we stay in bootloader, otherwise the
application is loaded (if it exists). However, unless the
GPIO pin is configured as an output and set high, its value
is random. Additionally, it takes a small amount of time
for the voltage to reach its set value after setting it
high.